+++ spec/www/spec/regexp-test.js
@@ -36,6 +36,7 @@ var mytests = function() {
 
       it(suiteName + 'Simple REGEXP test',
         function(done) {
+          if (!isWebSql && isAndroid && !isOldImpl) pending('BROKEN for Android-sqlite-connector'); // XXX TODO FIX
           if (isWebSql && !isAndroid) pending('BROKEN for iOS Web SQL');
           if (!isWebSql && isAndroid && isOldImpl && /Android [1-4]/.test(navigator.userAgent)) pending('BROKEN for android.database (version 1.x-4.x)');
           if (isWindows) pending('BROKEN for Windows');