+++ src/android/io/sqlc/SQLiteAndroidDatabase.java
@@ -145,6 +145,9 @@ class SQLiteAndroidDatabase
                             // do nothing here & try again with raw query.
                         }
 
+                        // "finally" cleanup myStatement
+                        myStatement.close();
+
                         if (rowsAffected != -1) {
                             queryResult = new JSONObject();
                             queryResult.put("rowsAffected", rowsAffected);
@@ -183,6 +186,9 @@ class SQLiteAndroidDatabase
                         errorMessage = ex.getMessage();
                         Log.v("executeSqlBatch", "SQLiteDatabase.executeInsert(): Error=" + errorMessage);
                     }
+
+                    // "finally" cleanup myStatement
+                    myStatement.close();
                 }
 
                 if (queryType == QueryType.begin) {