+++ CHANGES.md
@@ -1,5 +1,19 @@
 # Changes
 
+## 0.10.0-0x-xx-ext-dev
+
+XXX TBD
+
+### 0.10.0-0x-common-dev
+
+XXX TBD
+
+### cordova-sqlite-storage 1.2.1
+
+- Close Android SQLiteStatement after INSERT/UPDATE/DELETE
+- Specify minimum Cordova version 6.0.0
+- Lawnchair adapter fix: Changed remove method to work with key array
+
 ## 0.9.x-extras-xx-dev
 
 ### x.x.x-common-dev